[Improvement] Sign in with Apple Tests - Updated Sign in with Apple test cases
