--- tests/API/SIWASignUpTest.php
@@ -1,42 +1,0 @@
-<?php
-
-namespace Tests\API;
-
-use App\Session;
-use App\User;
-use Illuminate\Foundation\Testing\DatabaseMigrations;
-use Tests\TestCase;
-
-class SIWASignUpTest extends TestCase
-{
-    use DatabaseMigrations;
-
-    /**
-     * An account can be signed up.
-     *
-     * @return void
-     * @test
-     */
-    function an_account_can_be_signed_up_via_siwa()
-    {
-        $this->json('POST', '/api/v1/users/signup/siwa', [
-            'email'             => 'jdi8eji2@privaterelay.appleid.com',
-            'siwa_id'           => '001100.6a54b3ccop5041bba0773fd1079b0e22.2447',
-            'platform'          => 'iOS',
-            'platform_version'  => '13.4',
-            'device_vendor'     => 'Apple',
-            'device_model'      => 'iPhone 11 Pro Max'
-        ])->assertSuccessfulAPIResponse();
-
-        // Check whether the account was created
-        $this->assertEquals(1, User::count(), 'The user account was not created.');
-
-        // Check whether a session was created
-        $this->assertEquals(1, Session::count(), 'A session was not created.');
-
-        // Check whether the user can change their username
-        /** @var User $user */
-        $user = User::first();
-        $this->assertTrue($user->username_change_available, 'The user does not have a username change available.');
-    }
-}
